# apps/api/docs/openapi.yaml
components:
    schemas:
...
        handler.eventDTO:
            properties:
                category:
                    nullable: true
                    type: string
                community_id:
                    nullable: true
                    type: string
                cover_image_url:
                    nullable: true
                    type: string
                created_at:
                    nullable: true
                    type: string
                created_by:
                    nullable: true
                    type: string
                description:
                    nullable: true
                    type: string
                end_at:
                    nullable: true
                    type: string
                id:
                    nullable: true
                    type: string
                is_public:
                    nullable: true
                    type: boolean
                location:
                    nullable: true
                    type: string
                start_at:
                    nullable: true
                    type: string
                title:
                    nullable: true
                    type: string
                updated_at:
                    nullable: true
                    type: string
            type: object
...
        handler.rsvpRequest:
            properties:
                status:
                    type: string
            required:
                - status
            type: object
...
        handler.upsertEventRequest:
            properties:
                category:
                    type: string
                cover_image_url:
                    nullable: true
                    type: string
                description:
                    nullable: true
                    type: string
                end_at:
                    type: string
                is_public:
                    nullable: true
                    type: boolean
                location:
                    nullable: true
                    type: string
                start_at:
                    type: string
                title:
                    type: string
            required:
                - category
                - end_at
                - start_at
                - title
            type: object
...
        response.Envelope:
            properties:
                data:
                    nullable: true
                error:
                    $ref: '#/components/schemas/response.ErrorBody'
                meta:
                    $ref: '#/components/schemas/response.Meta'
                success:
                    nullable: true
                    type: boolean
            type: object
        response.ErrorBody:
            nullable: true
            properties:
                code:
                    $ref: '#/components/schemas/response.ErrorCode'
                fields:
                    additionalProperties:
                        type: string
                    nullable: true
                    type: object
                message:
                    nullable: true
                    type: string
            type: object
        response.ErrorCode:
            enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            nullable: true
            type: string
            x-enum-varnames:
                - ErrValidation
                - ErrUnauthorized
                - ErrForbidden
                - ErrNotFound
                - ErrConflict
                - ErrRateLimited
                - ErrInternal
        response.Meta:
            nullable: true
            properties:
                page:
                    nullable: true
                    type: integer
                per_page:
                    nullable: true
                    type: integer
                total:
                    nullable: true
                    type: integer
            type: object
    securitySchemes:
        BearerAuth:
            description: Clerk session JWT, sent as "Bearer <token>".
            in: header
            name: Authorization
            type: apiKey
info:
    contact: {}
    description: Backend API for the Iridescent community hub (events, memberships, announcements, gallery, achievements).
    title: Iridescent Hub API
    version: "1.0"
openapi: 3.0.3
paths:
...
    /events:
        get:
            parameters:
                - description: Page number
                  in: query
                  name: page
                  schema:
                    default: 1
                    type: integer
                - description: Items per page
                  in: query
                  name: per_page
                  schema:
                    default: 20
                    type: integer
                - description: Filter start_at >= (RFC3339)
                  in: query
                  name: from
                  schema:
                    type: string
                - description: Filter start_at <= (RFC3339)
                  in: query
                  name: to
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            items:
                                                $ref: '#/components/schemas/handler.eventDTO'
                                            type: array
                                      type: object
                    description: OK
            summary: List public events
            tags:
                - events
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/handler.upsertEventRequest'
                description: Event
                required: true
                x-originalParamName: body
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.eventDTO'
                                      type: object
                    description: Created
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unprocessable Entity
            security:
                - BearerAuth: []
            summary: Create an event
            tags:
                - events
    /events/{id}:
        delete:
            parameters:
                - description: Event ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Not Found
            security:
                - BearerAuth: []
            summary: Delete an event (soft delete)
            tags:
                - events
        get:
            description: Private events are hidden from anonymous callers (returned as 404).
            parameters:
                - description: Event ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.eventDTO'
                                      type: object
                    description: OK
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Not Found
            summary: Get a public event
            tags:
                - events
        patch:
            parameters:
                - description: Event ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/handler.upsertEventRequest'
                description: Event
                required: true
                x-originalParamName: body
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.eventDTO'
                                      type: object
                    description: OK
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unprocessable Entity
            security:
                - BearerAuth: []
            summary: Update an event
            tags:
                - events
    /events/{id}/rsvp:
        post:
            parameters:
                - description: Event ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/handler.rsvpRequest'
                description: RSVP status
                required: true
                x-originalParamName: body
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unauthorized
            security:
                - BearerAuth: []
            summary: RSVP to an event
            tags:
                - events
...
servers:
    - url: /api/v1